% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_simulation_with_conf.R
\name{ABC_simulation_with_conf}
\alias{ABC_simulation_with_conf}
\title{Do ABC simulation with configuration table}
\usage{
ABC_simulation_with_conf(
  npod,
  conf,
  time.range,
  buffer = 0,
  concentrationscale,
  concentrationShape,
  prefix = "temp",
  BayeSSCallocation,
  do.parallel = 1,
  write.reference.file = F
)
}
\arguments{
\item{npod}{An integer, the number of simulation replicates}

\item{conf}{A data frame with configurations}

\item{time.range}{A vector containing two numeric elements specifying the minimal and maximum event time.}

\item{buffer}{The minimal amount of time separating two events. The default is 0, when events' time are sampled randomly}

\item{concentrationscale}{A numeric. The \code{scale} for gamma distribution. See 'Details'}

\item{concentrationShape}{A numeric. The \code{shape} for gamma distribution. See 'Details'}

\item{prefix}{BayeSSC will generate some intermediate files, this function will create temporary folders for storing these files. One folder for one row in the configuration table, and the folder is named as \code{prefix}\code{simulation replicates}_row number.If folders with the same name already exist, user will be prompt for choosing another \code{prefix} or overwrite the folders.}

\item{BayeSSCallocation}{A character string providing the location of the BayeSSC executable (including the file name)}

\item{do.parallel}{A integer for the number of parallel threads to use. if equals \code{1},no parallel execution.}

\item{write.reference.file}{\code{TRUE} or \code{FALSE}(Default). If \code{TRUE}, the simulated hyperstat will be append to a file \code{prefix}_reference_table. Useful for running large number of replicates.If \code{False}, the simulated hyperstat will be returned as a data frame}
}
\value{
If \code{write.reference.file} is \code{FALSE}, this function returns a data frame containing the simulated hyperstats. The first column is 'uid', the replicates number, the second column is the number of events, and the third column is the total number of species. Following are columns for the "real" expansion time for each species in the simulation, and then, the columns of hyperstats.If \code{write.reference.file} is \code{TRUE}, this function returns a character string "finished" after finish all the replicates.
}
\description{
A function for doing abc simulation with a configuration table. see example in vignette for required columns in the table
}
\details{
For the prior on expansion event, hBayeSSC used a flat distribution where 1-nspecies expanded together are simulated with equal frequency. In the paper, we used a gamma distribution for the alpha parameter of the dirichelete process, similar to [PyMsbayes](http://joaks1.github.io/PyMsBayes/). User need to select the two parameters for gamma distribution: concentrationShape and consentrationScale. User can looked into the [prior selection in PyMsbayes](http://joaks1.github.io/PyMsBayes/tutorials/selecting-priors.html) for how to select these two parameters.
}
