% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_species_to_events.R
\name{assign_species_to_events}
\alias{assign_species_to_events}
\title{Assign species to events}
\usage{
assign_species_to_events(species, nco.events, even = F)
}
\arguments{
\item{species}{A vector of species names, integers are treated as characters}

\item{nco.events}{The number of co-expansion events}

\item{even}{The default is to assign species randomly. IF \code{TRUE}, will spread species evenly}
}
\value{
A list, the length of which is equal to the number of event. Each element of the list is a vector of species names assigned to the event
}
\description{
This function puts species into coexpansion events randomly or evenly
The number of co-expansion events is specified by the user
}
