% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hyperstat.R
\name{calculate_hyperstat}
\alias{calculate_hyperstat}
\title{Calculate hyperstats from observation stats}
\usage{
calculate_hyperstat(obs, hbayessclike = T, col = NULL)
}
\arguments{
\item{obs}{A data frame containing the BayeSSC simulated stats}

\item{hbayessclike}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, directly calculte the hyperstat across all locus in all species. If \code{FALSE}, still need develop}

\item{col}{A character vector containing the names of columns for hyperstat calculation. Currently this parameter has no use}
}
\value{
A data frame containing the calculated hyperstats
}
\description{
A function for calculating hyperstats from BayeSSC simulated stats. Here we calculate four hyperstats for each summary statistics: mean, variance, skewness and kurtosis as explained on https://en.wikipedia.org/wiki/Algorithms_for_calculating_variance#Higher-order_statistics
}
