% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_conf.R
\name{edit_conf}
\alias{edit_conf}
\title{Sampling the configuration table for each species}
\usage{
edit_conf(conf, col = c("gen", "Ne", "popratio"))
}
\arguments{
\item{conf}{A data frame with configurations}

\item{col}{A character vector containing the columns for which the value need to be sampled from distributions. The default contains three columns in the \code{conf} data frame: "gen"(generation time), "Ne"(effective population size), and "popratio"(the population expansion ratio)}
}
\value{
A configuration data frame with randomly sampled values
}
\description{
A function for randomly draw values from distributions included in the configuration table. BayeSSC can take prior distribution for parameters instead of fixed values. However, loci of the same species need to share the same generation time, effective population size and expansion history. So these values need to be fixed for each species instead of letting BayeSSC to draw a difference sample for each locus.
}
