% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coevent_number_dirichlet.R
\name{generate_coevent_number_dirichlet}
\alias{generate_coevent_number_dirichlet}
\title{Generate the number of events according to the dirichlet process}
\usage{
generate_coevent_number_dirichlet(species, alpha, maxevent = NA)
}
\arguments{
\item{species}{A vector of species names}

\item{alpha}{The parameter alpha for the dirichlet process}

\item{maxevent}{The default is NA. User can specified the maximum number of event.}
}
\value{
A list, the length of which is equal to the number of event. Each element of the list is a vector of species names assigned to the event
}
\description{
This function puts species into coexpansion events according to dirichlet process.
It will go through the list of species one by one, and determine whether each one should
be assigned to existing events or to a new event. The probability determined by the alpha
parameter.
}
