% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_msreject.R
\name{run_msreject}
\alias{run_msreject}
\title{Run MsReject}
\usage{
run_msreject(
  hyperstat,
  reference.table,
  MsRejectallocation,
  samplingtolerance,
  prefix = "temp",
  hbayeSSC.style = T,
  write.posterior.file = F
)
}
\arguments{
\item{hyperstat}{A vector containing "observed" hyperstat}

\item{reference.table}{A data frame containing the hyperstat from ABC simulation. The column name of this data frame must match the \code{hyperstat} vector name}

\item{MsRejectallocation}{The location of the msreject (including the executable itself)}

\item{samplingtolerance}{Numeric. The sampling fraction for msreject}

\item{prefix}{a character string. For writing the "observed" hyperstat and reference table to files, \code{prefix}_hyperstat_file and \code{prefix}_reference_table, respectively}

\item{hbayeSSC.style}{\code{TRUE}(Default) or \code{FALSE}.hBayeSSC uses the hyperstat calculated with (the number of)haptypes, hapdiv, nucdiv, tajimasD.}

\item{write.posterior.file}{\code{TRUE} or \code{FALSE}(Default). If \code{TRUE}, the retained abc simulation replicates will be write to a file \code{prefix}_Posterior. If \code{False}, they will be returned as a data frame}
}
\value{
If \code{write.posterior.file} is \code{FALSE}, this function returns a data frame containing the retained simulated hyperstats after rejection. The first column is 'uid', the replicates number, the second column is the number of events, and the third column is the total number of species. Following are columns for the "real" expansion time for each species in the simulation, and then, the columns of hyperstats.If \code{write.posterior.file} is \code{TRUE}, this function returns the filename (including the path) of the posterior file.
}
\description{
This is wrapper function for running Msreject.For installing/compiling msReject see [msbayes webpage](http://msbayes.sourceforge.net/).
}
