% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_exp_time.R
\name{species_exp_time}
\alias{species_exp_time}
\title{Get expansion time for each species}
\usage{
species_exp_time(species.assignment, exp.time)
}
\arguments{
\item{species.assignment}{a list showing the species-to-event assignment}

\item{exp.time}{The time of the co-expansion events}
}
\value{
A numeric vector containing the expansion time for each species.
}
\description{
A short function for get expansion time for each species
}
\seealso{
\code{\link[CoExpansionValidation]{assign_species_to_events}}
}
