% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cotime_with_buffer.R
\name{generate_cotime_with_buffer}
\alias{generate_cotime_with_buffer}
\title{Generate co-expansion time}
\usage{
generate_cotime_with_buffer(time.range, nco.events, buffer = 0)
}
\arguments{
\item{time.range}{A vector containing two numeric elements specifying the minimal and maximum event time.}

\item{nco.events}{The number of coexpansion events. See 'Details'.}

\item{buffer}{The minimal amount of time separating two events. The default is 0, when events' time are sampled randomly}
}
\value{
A numeric vector of the sampled expansion time
}
\description{
This function randomly draw events' time from a time range.
User can specify the minimal time length (buffer) between events
}
\details{
Note that the \code{buffer} is the time length on each side of a expansion time. Also, if the \code{buffer} is non-zero,the maximum number of events (max \code{nco.event})=\code{time.range}/\code{buffer} +2. However, if \code{nco.events} is close to the it maximum (even if not exceeding), this function might take a while to find a set of events time that are at least buffer apart from each other.
}
