% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_to_config.R
\name{par_to_config}
\alias{par_to_config}
\title{Extract information from a par file}
\usage{
par_to_config(bayessc_par_file, species, nloci, gen = NA)
}
\arguments{
\item{bayessc_par_file}{A character string providing the path to the .par file}

\item{species}{A character vector containing the species names}

\item{nloci}{A integer. The number of loci to simulate for each species}

\item{gen}{A vector of integers containing the generation time for each species. If it is of length 1, function will assume the same generation time for all species. If it is \code{NA}(Default), assuming one generation per time unit.}
}
\value{
A configuration data frame
}
\description{
A function extracting information from a par file and put it into a configuration data frame. Note that because we only consider one historical expansion event, the par file format is semi-fixed, and a template par file is included in the package for using as a template.
}
