% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_msreject.R
\name{run_msreject}
\alias{run_msreject}
\title{Run MsReject}
\usage{
run_msreject(
  hyperstat,
  reference.table,
  MsRejectallocation,
  samplingtolerance,
  prefix = "temp",
  hbayeSSC.style = T,
  write.posterior.file = F
)
}
\arguments{
\item{hyperstat}{A vector containing the "observed" hyperstats from step 1}

\item{reference.table}{A data frame containing the hyperstats from ABC simulation(step 2). The column name of this data frame must match the names of \code{hyperstat} vector}

\item{MsRejectallocation}{A character string. The location of the msReject (including the executable itself)}

\item{samplingtolerance}{Numeric. The sampling fraction for msReject}

\item{prefix}{A character string for writing the "observed" hyperstat and reference table to files, \code{prefix}_hyperstat_file and \code{prefix}_reference_table, respectively}

\item{hbayeSSC.style}{\code{TRUE}(Default) or \code{FALSE}.In hBayeSSC, running msReject involves using the hyperstats of haptypes, hapdiv, nucdiv, tajimasD.}

\item{write.posterior.file}{\code{TRUE} or \code{FALSE}(Default). If \code{TRUE}, the retained abc simulation replicates will be write to a file named \code{prefix}_Posterior. If \code{False}, they will be returned as a data frame}
}
\value{
If \code{write.posterior.file} is \code{FALSE}, this function returns a data frame containing the retained rows from reference table after rejection. The first column is 'uid', the replicates number, the second column is the number of events, and the third column is the total number of species. Following are columns for the "real" expansion time for each species in the simulation, and then, the columns of hyperstats.If \code{write.posterior.file} is \code{TRUE}, this function returns the file name (including the path) of the posterior file.
}
\description{
This is wrapper function for running the msReject in msbayes.For installing/compiling msReject see [msbayes webpage](http://msbayes.sourceforge.net/).
}
